src/services/consumer/requirements.txt
fastavro==1.9.7
kafka-python==2.0.2
